pr-monitoring/pyproject.toml
[project]
name = "pr-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
]
[tool.uv]
package = true

[project.scripts]
pr-monitoring = "pr_monitoring:main"
